Fix tapis_jobs tests and re-enable module
