+++ libs/shared/src/infra/reconciliation/clients/mod.rs
@@ -1,2 +1,2 @@
-// pub mod tapis_jobs;
+pub mod tapis_jobs;
 pub mod tapis_pods;
